Refactor SearchForm component into its own file
